requirements.txt
appdirs==1.4.3
Django==1.10
django-pure-pagination==0.3.0
django-widget-tweaks==1.4.1
djangorestframework==3.4.4
packaging==16.8
Pillow==3.3.1
psycopg2==2.6.2
pyparsing==2.2.0
six==1.10.0
